Split the discrimination probe out of checks.ts (F-1126)
`lint-code-health.mjs` caught checks.ts at 550 lines against a 500 LOC
limit — a gate I had not run locally, which is the actual defect here: the
cheap-gate block has ten scripts and I had run three.

The gate offers a `// file-size:` escape hatch and splitting. Splitting is
right, because the seam it points at is real: everything left in checks.ts
DEFINES the grammar a declaration is written in, and the probe EXERCISES a
declaration against the worlds it named. The file-size limit found a
boundary rather than merely a length.

`probeDiscrimination` is re-exported from checks.ts, so
`@pome-sh/sdk/checks` keeps an identical public surface and the split stays
internal. The new module imports only TYPES from checks.ts, so there is no
runtime cycle.

checks.ts 550 -> 474; check-discrimination.ts 96.
